backend/requirements.txt
# FastAPI Backend
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
deep-translator>=1.11.0

# Audio Processing (DSP)
numpy>=1.24.0
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.11.0
pydub>=0.25.0

# Visualization
matplotlib>=3.7.0

# Speech Recognition & TTS
SpeechRecognition>=3.10.0
gTTS>=2.4.0
python-dotenv>=1.0.0
requests>=2.31.0

# Optional: Streamlit (for streamlit_app.py demo)
# streamlit>=1.28.0
# streamlit-audiorec>=0.1.3

# Translation (optional)
# googletrans==4.0.0-rc1
